+++ dsl-core/io.mdsl.ui/src/io/mdsl/ui/handler/AsyncAPIGenerationHandler.java
@@ -1,0 +1,22 @@
+package io.mdsl.ui.handler;
+
+import org.eclipse.core.commands.ExecutionEvent;
+import org.eclipse.xtext.generator.IGenerator2;
+import com.google.inject.Inject;
+import io.mdsl.generator.asyncapi.AsyncApiGenerator;
+
+public class AsyncAPIGenerationHandler extends AbstractGenerationHandler {
+
+	@Inject
+	private AsyncApiGenerator generator;
+
+	@Override
+	protected IGenerator2 getGenerator() {
+		return generator;
+	}
+
+	@Override
+	protected void postGeneration(ExecutionEvent event) {
+		
+	}
+}